requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
aiohttp==3.9.1
python-dotenv==1.0.0
opencv-python==4.8.1.78

pillow==10.1.0
numpy==1.24.3
